requirements.txt
streamlit
requests
google-generativeai
yt-dlp